bot/handlers/wallet.py
"""
Wallet tracking command handlers.
"""

from telegram import Update, InlineKeyboardButton, InlineKeyboardMarkup
from telegram.ext import ContextTypes
from solders.pubkey import Pubkey


async def wallet_command(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Handle /wallet command."""
    if not context.args:
        await update.message.reply_text(
            "📊 *Wallet Tracker*\n\n"
            "Usage: `/wallet <address>`\n\n"
            "Example:\n"
            "`/wallet 7xKXtg2CW87d97TXJSDpbD5jBkheTqA83TZRuJosgAsU`",
            parse_mode='Markdown'
        )
        return
    
    wallet_address = context.args[0]
    
    # Validate address
    try:
        Pubkey.from_string(wallet_address)
    except Exception:
        await update.message.reply_text(
            "❌ Invalid Solana wallet address.\n\n"
            "Please provide a valid base58 address."
        )
        return
    
    # Fetch wallet info (mock data for demonstration)
    wallet_info = await get_wallet_info(wallet_address)
    
    # Build response message
    message = (
        f"🐋 *Wallet Analysis*\n\n"
        f"Address: `{wallet_address[:8]}...{wallet_address[-8:]}`\n"
        f"Balance: *{wallet_info['balance']:.2f} SOL*\n"
        f"Risk Score: *{wallet_info['risk_score']:.2f}/1.00*\n"
        f"Total Transactions: *{wallet_info['tx_count']:,}*\n"
        f"First Seen: {wallet_info['first_seen']}\n"
        f"Last Activity: {wallet_info['last_activity']}\n\n"
        f"Tags: {', '.join(wallet_info['tags'])}\n"
    )
    
    # Add track button
    keyboard = [[
        InlineKeyboardButton("🔔 Track This Wallet", callback_data=f"track_{wallet_address}")
    ]]
    reply_markup = InlineKeyboardMarkup(keyboard)
    
    await update.message.reply_text(
        message,
        parse_mode='Markdown',
        reply_markup=reply_markup
    )


async def track_wallet(update: Update, context: ContextTypes.DEFAULT_TYPE):
    """Handle wallet tracking button."""
    query = update.callback_query
    await query.answer()
    
    wallet_address = query.data.replace("track_", "")
    
    # Add wallet to user's tracking list (mock implementation)
    # In production, this would store in database
    
    await query.edit_message_text(
        f"✅ Now tracking wallet `{wallet_address[:8]}...{wallet_address[-8:]}`\n\n"
        f"You'll receive alerts for:\n"
        f"• Transactions > 10,000 SOL\n"
        f"• New token purchases\n"
        f"• Large transfers\n\n"
        f"Manage tracking: /stats",
        parse_mode='Markdown'
    )


async def get_wallet_info(address: str) -> dict:
    """
    Fetch wallet information from CipherX API.
    
    In production, this would call the actual API.
    For demonstration, returns mock data.
    """
    # Mock data
    return {
        'balance': 1234.56,
        'risk_score': 0.23,
        'tx_count': 4821,
        'first_seen': '2024-01-15',
        'last_activity': '2025-11-01',
        'tags': ['🐋 Whale', '⚡ High Activity', '✅ Verified']
    }
