bot/requirements.txt
python-telegram-bot==20.6
aiohttp==3.9.0
asyncpg==0.29.0
redis==5.0.1
solders==0.19.0
solana==0.30.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.3
numpy==1.26.2
scikit-learn==1.3.2
